remove properties & refactor utils
