+++ device_doctor/lib/src/device.dart
@@ -7,25 +7,14 @@ import 'dart:async';
 import 'android_device.dart';
 import 'health.dart';
 import 'ios_device.dart';
-import 'utils.dart';
 
 /// Discovers available devices and chooses one to work with.
 abstract class DeviceDiscovery {
   factory DeviceDiscovery(String deviceOs) {
     switch (deviceOs) {
       case 'ios':
-        for (String property in supportedIosProperties) {
-          if (!properties.keys.contains(property)) {
-            throw FormatException('\n-----\n$property is not defined for $deviceOs device.\n-----');
-          }
-        }
         return IosDeviceDiscovery();
       case 'android':
-        for (String property in supportedAndroidProperties) {
-          if (!properties.keys.contains(property)) {
-            throw FormatException('\n-----\n$property is not defined for $deviceOs device.\n-----');
-          }
-        }
         return AndroidDeviceDiscovery();
       default:
         throw StateError('Unsupported device operating system: $deviceOs');
@@ -33,7 +22,7 @@ abstract class DeviceDiscovery {
   }
 
   /// Lists all available devices' IDs.
-  Future<List<Device>> discoverDevices({Duration retriesDelay = const Duration(seconds: 10)});
+  Future<List<Device>> discoverDevices({Duration retriesDelaySeconds = const Duration(seconds: 10)});
 
   /// Checks the health of the available devices.
   Future<Map<String, List<HealthCheckResult>>> checkDevices();
@@ -46,12 +35,6 @@ abstract class DeviceDiscovery {
 abstract class Device {
   /// A unique device identifier.
   String get deviceId;
-
-  /// Whether the device is awake.
-  Future<bool> isAwake();
-
-  /// Whether the device is asleep.
-  Future<bool> isAsleep();
 
   /// Recovers the device back to a healthy state.
   Future<void> recover();
